requirements.txt
# Cattle Breed Recognition - Requirements

# Core ML/DL dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=9.0.0

# Visualization and plotting
matplotlib>=3.5.0
seaborn>=0.11.0
opencv-python>=4.5.0
plotly>=5.0.0

# Data manipulation and analysis
pandas>=1.3.0
scikit-learn>=1.0.0

# Web interface
streamlit>=1.20.0

# Utilities
tqdm>=4.60.0
pathlib2>=2.3.0

# Optional: For GPU support (install separately if needed)
# Install PyTorch with CUDA from: https://pytorch.org/get-started/locally/